packages/gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=12.0
supabase>=2.24.0
web3==6.15.0
httpx>=0.27.0
PyJWT>=2.8.0
pydantic>=2.11.7
python-multipart>=0.0.20
python-dotenv>=1.0.0
python-telegram-bot>=21.0
resend>=2.0
bcrypt>=4.0
